Add login container
Handle update and submit form
